Separated into multiple files, applied @TrueDoctor review feedback
