+++ editor/src/viewport_tools/mod.rs
@@ -1,6 +1,6 @@
-pub mod shape_manipulation;
 pub mod snapping;
 pub mod tool;
 pub mod tool_message;
 pub mod tool_message_handler;
 pub mod tools;
+pub mod vector_editor;
